pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-customer_support_ops_env"
version = "0.1.0"
description = "Deterministic customer support operations environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m customer_support_ops_env.server.app
server = "customer_support_ops_env.server.app:main"
baseline = "customer_support_ops_env.baseline:main"

[tool.setuptools]
include-package-data = true
packages = [
    "customer_support_ops_env",
    "customer_support_ops_env.server",
    "customer_support_ops_env.graders",
]
